feat(acked): add a non consuming method to the notice
This will make it possible to not block while trying to wait for the
acknowledgement of multiple packets. For example, you are storing the
`NoticeFutures` in a `Vec`, and you want to find the first acknowledge packet.

This would be similar to the `try_recv(&mut)` of the inner channel.